requirements.txt
click
colorama
pbr
ruamel.yaml
sh
